+++ src/pages/RootPage.tsx
@@ -1,10 +1,8 @@
-import type { PropsWithChildren } from 'react';
-import { useState } from 'react';
-import { Outlet, useLocation, useNavigate } from 'react-router-dom';
+import { Outlet, useNavigate } from 'react-router-dom';
 import { styled } from 'styled-components';
 import AppBar from '../components/page/AppBar';
-import ResponseErrorBoundary from '../components/utils/ResponseErrorBoundary';
-import LoginPage from './LoginPage';
+import NotFoundErrorBoundary from './boundaries/NotFoundErrorBoundary';
+import UnauthorizedErrorBoundary from './boundaries/UnauthorizedErrorBoundary';
 
 const Content = styled.main`
   margin: 0 auto;
@@ -13,25 +11,6 @@ const Content = styled.main`
 
   max-width: 1300px;
 `;
-
-type UnauthorizedErrorBoundaryProps = PropsWithChildren;
-
-const UnauthorizedErrorBoundary = (props: UnauthorizedErrorBoundaryProps) => {
-  const { children } = props;
-  const location = useLocation();
-  const [key, setKey] = useState<number>(0);
-  const rerender = () => setKey(Date.now());
-
-  return (
-    <ResponseErrorBoundary
-      key={location.pathname + key}
-      catches={(response) => response.accept(401)}
-      fallback={<LoginPage onLoginSuccess={rerender} />}
-    >
-      {children}
-    </ResponseErrorBoundary>
-  );
-};
 
 const RootPage = () => {
   const navigate = useNavigate();
@@ -42,7 +21,9 @@ const RootPage = () => {
 
       <Content>
         <UnauthorizedErrorBoundary>
-          <Outlet />
+          <NotFoundErrorBoundary>
+            <Outlet />
+          </NotFoundErrorBoundary>
         </UnauthorizedErrorBoundary>
       </Content>
     </>
